Updated based on feedback
Updated:

the height and width of the video in macOS onboarding videos for mac and windows users.

moved the note for linking Apple ID to device to the prereqs.

Moved the device clean up policy faq to the FAQ section.

renamed Troubleshooting FAQS as Known issues and FAQs in the side bar.